infra/runner-autoscaler/requirements.txt
azure-functions==1.21.3
azure-identity==1.19.0
